fix /0 IP mask being treated as /32 or /128
Checks "if not mask:" would return True if mask was set to 0.
This would lead to mask being set to default value, which is /32 for
IPv4 and /128 for IPv6, which is not the intended behaviour, since
it prevented explicitly specifying full IP range (0.0.0.0/0).
The mask would also not show up in dashboard.

This commit fixes the issue by checking the mask values for equality
to None.